Make Graph interface in models, move Graph struct to blocks
