+++ blocks/graph.go
@@ -1,0 +1,76 @@
+package blocks
+
+import (
+	"encoding/json"
+	"fmt"
+
+	"github.com/jamestunnell/marketanalysis/models"
+)
+
+type Graph struct {
+	Blocks      models.Blocks
+	Connections models.Connections
+}
+
+type GraphJSON struct {
+	Blocks      map[string]json.RawMessage `json:"blocks"`
+	Connections models.Connections         `json:"connections"`
+}
+
+func NewGraph() *Graph {
+	return &Graph{
+		Blocks:      models.Blocks{},
+		Connections: models.Connections{},
+	}
+}
+
+func (g *Graph) GetBlocks() models.Blocks {
+	return g.Blocks
+}
+
+func (g *Graph) GetConnections() models.Connections {
+	return g.Connections
+}
+
+func (g *Graph) MarshalJSON() ([]byte, error) {
+	j := &GraphJSON{
+		Blocks:      map[string]json.RawMessage{},
+		Connections: g.Connections,
+	}
+
+	for name, b := range g.Blocks {
+		d, err := MarshalBlockJSON(b)
+		if err != nil {
+			return []byte{}, fmt.Errorf("failed to marshal block %s: %w", name, err)
+		}
+
+		j.Blocks[name] = d
+	}
+
+	return json.Marshal(j)
+}
+
+func UnmarshalGraphJSON(d []byte) (*Graph, error) {
+	var j GraphJSON
+
+	if err := json.Unmarshal(d, &j); err != nil {
+		return nil, fmt.Errorf("failed to unmarshal as graph: %w", err)
+	}
+
+	blocks := models.Blocks{}
+	for name, blockData := range j.Blocks {
+		b, err := UnmarshalBlockJSON(blockData)
+		if err != nil {
+			return nil, fmt.Errorf("failed to unmarshal block %s: %w", name, err)
+		}
+
+		blocks[name] = b
+	}
+
+	g := &Graph{
+		Blocks:      blocks,
+		Connections: j.Connections,
+	}
+
+	return g, nil
+}
